+++ Assets/Scripts/GT/Counters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 955533bcd1114ca4b93d0d0b53e025d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
