+++ Assets/Scripts/GT/Counters/CloudChart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb7b9274732127742bfaee8345eee937
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
