+++ Assets/Scripts/GT/Counters/ECloudChartStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 18190c488b2e2514a88a5be998ec20fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
