+++ Assets/Scripts/GT/Counters/GoldenTime.cs
