+++ Assets/Scripts/GT/Counters/GoldenTime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b5862f02da42d6c4d9ef20e03e8ccf7c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
