+++ Assets/Scripts/GT/Counters/Scoreboard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79c380b1839ef62438e4dab16b43ae86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
