+++ Assets/Scripts/GT/Events/Concrete/BullyEvent.cs
@@ -1,12 +1,14 @@
+using System;
+
 namespace GT.Events.Concrete
 {
-    public class RoutineBagCheckEvent : Event
+    public class BullyEvent : Event
     {
         // TODO: random prompt generation from file
-        public RoutineBagCheckEvent(Random rng) : base(rng, RegularDailyThreshold, string.Empty)
+        public BullyEvent(Random rng) : base(rng, RegularDailyThreshold, string.Empty)
         {
         }
-
+        
         public override void Result(Game game)
         {
             throw new NotImplementedException();
