+++ Assets/Scripts/GT/Events/Concrete/BullyEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb894eeff2bf9194a8a24b39e6762d8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
