+++ Assets/Scripts/GT/Events/Concrete/NoLunchEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 270fa60fd9536a048b4e46ec78f7393c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
