+++ Assets/Scripts/GT/Events/Concrete/RainyDayEvent.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace GT.Events.Concrete
 {
     public class RainyDayEvent : Event
