+++ Assets/Scripts/GT/Events/Concrete/RainyDayEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44aa511684ddde74984432f5ac24422c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
