+++ Assets/Scripts/GT/Events/Concrete/RandomDeathEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b71eaa316d9a4b54b95e11c1b46cd93a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
