+++ Assets/Scripts/GT/Events/Concrete/RoutineBagCheckEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2dcb14b2eea6cee4c863ad267acec90e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
