+++ Assets/Scripts/GT/Events/Event.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8acd41ef2bb24814bb58c1568ad3fffe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
