+++ Assets/Scripts/GT/Events/EventFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f85ca45b1428b6e4388caa4255e2315c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
