+++ Assets/Scripts/GT/Events/IEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b08722be601dd2846b6ee85e1aa378e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
