+++ Assets/Scripts/GT/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b31cb1606e49764baf569f32cbc3d94
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
