+++ Assets/Scripts/GT/Items/Cards.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a3a81a3383837d4bbc282400899eba3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
