+++ Assets/Scripts/GT/Items/Cards/Card.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 74f34c2a79e950b4ca5293ddc3c10f72
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
