+++ Assets/Scripts/GT/Items/Cards/Deck.cs
@@ -1,3 +1,7 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
 namespace GT.Items.Cards
 {
     public class Deck
