+++ Assets/Scripts/GT/Items/Cards/Deck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a37ee03a80f25934f9a061d457135c43
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
