+++ Assets/Scripts/GT/Items/Cards/ECardValue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ca3990e04ad3ef40a60d477e01e47bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
