+++ Assets/Scripts/GT/Quests.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1f1483d8cac442249aabaf90752edee5
+timeCreated: 1706305630
