+++ Assets/Scripts/GT/Quests/IQuest.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: fc4c70a1b1cc4ae0916bb786a47c1d9d
+timeCreated: 1706305859
