+++ Assets/Scripts/GT/Quests/Quest.cs
@@ -1,0 +1,26 @@
+namespace GT.Quests
+{
+    public class Quest : IQuest
+    {
+        /*
+         * 
+         */
+
+        public Quest(string description /* Tasks, Rewards, Penalties */)
+        {
+            
+        }
+        
+        public string GetDescription()
+        {
+            return null;
+        }
+
+        public void CompleteQuest()
+        {
+            
+        }
+
+        
+    }
+}
