+++ Assets/Scripts/GT/Quests/Quest.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a502bc39cb314aef8f1b1636c0419039
+timeCreated: 1706305719
